+++ CHANGES.md
@@ -20,6 +20,7 @@ Fixed:
 - Fixed duplicate metadata insertion in caching mode (#1241).
 - Fixed SSL connection error (#1304).
 - Better error reporting for Clock_conflict and Clock_loop errors (#985).
+- Make sure playlist change watch is only setup once (#1246).
 
 1.4.2 (03-05-2020)
 =====